# specs/001-create-in-examples/contracts/api.yaml
openapi: 3.0.3
info:
  title: TanStack Start API
  description: API specification for TanStack Start application with CloudFront and Lambda
  version: 1.0.0
  contact:
    name: TanStack Start API
    url: https://tanstack.com/start
servers:
  - url: https://{domain}/api
    description: CloudFront distribution
    variables:
      domain:
        default: example.cloudfront.net
        description: CloudFront domain name
paths:
...
  /auth/login:
    post:
      summary: User authentication
      description: Authenticate user and create session
      operationId: authenticateUser
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                rememberMe:
                  type: boolean
                  default: false
              required: [email, password]
      responses:
        "200":
          description: Authentication successful
          headers:
            Set-Cookie:
              description: Session cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
                  sessionId:
                    type: string
                    format: uuid
                required: [user, sessionId]
        "401":
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        lastLoginAt:
          type: string
          format: date-time
      required: [id, email, name, createdAt]
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          format: uuid
      required: [error, message, timestamp, requestId]
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session-token
      description: Session-based authentication using secure HTTP-only cookies
security:
  - sessionAuth: []
...